+++ internal/app/components/courses/courses.go
