+++ internal/app/components/degrees/degrees_test.go
