+++ internal/app/components/terms/terms.go
