+++ internal/app/components/terms/terms_test.go
