+++ internal/app/components/utils/mapUtils.go
